minor changes to code

+++ Data.py
@@ -4,7 +4,7 @@ class Data:
 
     def __init__(self, resolution):
         self.resolution = resolution
-        self.wMtx = np.zeros((resolution, resolution)).astype(int)
+        self.wMtx = np.zeros((resolution, resolution)).astype(float)
         self.flagMtx = np.ones((resolution, resolution)).astype(int)
         self.pillarCoords = []
         self.pillarIndex = []
